refactor: split mixed CAD enums by ownership
