+++ src/CADShared/Cad/Database/Entities/Text/FontTTF.cs
@@ -1,0 +1,33 @@
+namespace Fs.Fox.Cad;
+
+/// <summary>
+/// ttf字体枚举
+/// </summary>
+public enum FontTTF
+{
+    /// <summary>
+    /// 宋体
+    /// </summary>
+    [Description("宋体.ttf")]
+    宋体,
+    /// <summary>
+    /// 仿宋
+    /// </summary>
+    [Description("simfang.ttf")]
+    仿宋,
+    /// <summary>
+    /// 仿宋GB2312
+    /// </summary>
+    [Description("FSGB2312.ttf")]
+    仿宋GB2312,
+    /// <summary>
+    /// Arial
+    /// </summary>
+    [Description("Arial.ttf")]
+    Arial,
+    /// <summary>
+    /// Romans
+    /// </summary>
+    [Description("Romans")]
+    Romans
+}
