+++ src/CADShared/Cad/Database/Xrefs/PathConverterModes.cs
@@ -1,0 +1,16 @@
+namespace Fs.Fox.Cad;
+
+/// <summary>
+/// 参照路径转换
+/// </summary>
+public enum PathConverterModes : byte
+{
+    /// <summary>
+    /// 相对路径
+    /// </summary>
+    Relative,
+    /// <summary>
+    /// 绝对路径
+    /// </summary>
+    Complete
+}
