+++ src/CADShared/Cad/Geometry/OrientationType.cs
@@ -1,0 +1,20 @@
+namespace Fs.Fox.Cad;
+
+/// <summary>
+/// 方向的枚举
+/// </summary>
+public enum OrientationType
+{
+    /// <summary>
+    /// 左转或逆时针
+    /// </summary>
+    CounterClockWise,
+    /// <summary>
+    /// 右转或顺时针
+    /// </summary>
+    ClockWise,
+    /// <summary>
+    /// 重合或平行
+    /// </summary>
+    Parallel
+}
